refactor - move alarm api into user package
