+++ src/main/java/com/save_backend/src/auth/AuthDao.java
@@ -1,7 +1,5 @@
 package com.save_backend.src.auth;
 
-import com.save_backend.src.auth.model.GetAlarmRes;
-import com.save_backend.src.auth.model.PatchAlarmRes;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -17,34 +15,7 @@ public class AuthDao {
         this.jdbcTemplate = new JdbcTemplate(dataSource);
     }
 
-    public GetAlarmRes getAlarm(int userIdx) {
-        String getAlarmQuery = "SELECT is_alarm FROM user WHERE user_idx = ?";
-        int getAlarmParam = userIdx;
 
-        return this.jdbcTemplate.queryForObject(getAlarmQuery,
-                (rs, rowNum) -> new GetAlarmRes(
-                        userIdx,
-                        rs.getBoolean("is_alarm")
-                ), getAlarmParam);
-    }
-    public PatchAlarmRes changeAlarm(int userIdx) {
-        String changeDaoQuery = "UPDATE user\n" +
-                "SET user.is_alarm =\n" +
-                "    IF((SELECT is_alarm FROM(SELECT u.is_alarm FROM user as u WHERE u.user_idx = ?) tmp)= true, false, true)\n" +
-                "WHERE user_idx = ?;";
-        Object[] changeDaoParam = new Object[]{userIdx,userIdx};
-        this.jdbcTemplate.update(changeDaoQuery, changeDaoParam);
-
-        String getAlarmQuery = "SELECT is_alarm FROM user WHERE user_idx = ?";
-        int getAlarmParam = userIdx;
-
-        PatchAlarmRes patchAlarmRes = this.jdbcTemplate.queryForObject(getAlarmQuery,
-                (rs, rowNum) -> new PatchAlarmRes(
-                        userIdx,
-                        rs.getBoolean("is_alarm")
-                ), getAlarmParam);
-        return patchAlarmRes;
-    }
 
     /**
      * validation
