+++ src/main/java/com/save_backend/src/auth/AuthProvider.java
@@ -1,11 +1,9 @@
 package com.save_backend.src.auth;
 
 import com.save_backend.config.exception.BaseException;
-import com.save_backend.src.auth.model.GetAlarmRes;
 import org.springframework.stereotype.Service;
 
 import static com.save_backend.config.response.BaseResponseStatus.DATABASE_ERROR;
-import static com.save_backend.config.response.BaseResponseStatus.NOT_EXIST_USER;
 
 @Service
 public class AuthProvider {
@@ -13,17 +11,6 @@ public class AuthProvider {
 
     public AuthProvider(AuthDao authDao) {
         this.authDao = authDao;
-    }
-
-    public GetAlarmRes getAlarm(int userIdx) throws BaseException {
-        if(this.checkUser(userIdx)==0){
-            throw new BaseException(NOT_EXIST_USER);
-        }
-        try{
-            return authDao.getAlarm(userIdx);
-        } catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
     }
 
     /**
