+++ src/main/java/com/save_backend/src/user/model/PatchAlarmRes.java
@@ -1,4 +1,4 @@
-package com.save_backend.src.auth.model;
+package com.save_backend.src.user.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class GetAlarmRes {
+public class PatchAlarmRes {
     private int userIdx;
     private Boolean isAlarm;
 }
